TestUtilCleanup - #2 Move few methods from Flink sink test utils to common Flink connector tests utils.
