+++ flink/src/test/java/io/delta/flink/source/DeltaSourceBoundedExecutionITCaseTest.java
@@ -11,6 +11,7 @@ import java.util.stream.Collectors;
 import io.delta.flink.source.ContinuousTestDescriptor.Descriptor;
 import io.delta.flink.source.RecordCounterToFail.FailCheck;
 import io.delta.flink.utils.DeltaTestUtils;
+import io.delta.flink.utils.FailoverType;
 import org.apache.flink.core.fs.Path;
 import org.apache.flink.table.data.RowData;
 import org.apache.flink.table.types.logical.RowType;
@@ -23,6 +24,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EnumSource;
+import static io.delta.flink.utils.ExecutionITCaseTestConstants.*;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.IsEqual.equalTo;
 
@@ -164,8 +166,8 @@ public class DeltaSourceBoundedExecutionITCaseTest extends DeltaSourceITBase {
     }
 
     private void shouldReadDeltaTable(
-        DeltaSource<RowData> deltaSource,
-        FailoverType failoverType) throws Exception {
+            DeltaSource<RowData> deltaSource,
+            FailoverType failoverType) throws Exception {
         // WHEN
         // Fail TaskManager or JobManager after half of the records or do not fail anything if
         // FailoverType.NONE.
